+++ src/Himari/Prelude/Arrow.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE Safe #-}
+{-# LANGUAGE NoDerivingVia #-}
+{-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
+{-# LANGUAGE NoTemplateHaskell #-}
+
 -- | "Control.Arrow" re-exports, hiding symbols that conflict with "Data.Bifunctor".
 module Himari.Prelude.Arrow
   ( module Export