+++ src/Himari/Prelude/Data.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE Safe #-}
+{-# LANGUAGE NoDerivingVia #-}
+{-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
+{-# LANGUAGE NoTemplateHaskell #-}
+
 -- | "Data.Data" re-exports, hiding symbols that conflict with "GHC.Generics".
 module Himari.Prelude.Data
   ( module Export