+++ src/Himari/Prelude/FilePath.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE Safe #-}
+{-# LANGUAGE NoDerivingVia #-}
+{-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
+{-# LANGUAGE NoTemplateHaskell #-}
+
 -- | "System.FilePath" re-exports, hiding symbols that conflict with "Control.Lens".
 module Himari.Prelude.FilePath
   ( module Export