+++ src/Himari/Prelude/Monoid.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE Safe #-}
+{-# LANGUAGE NoDerivingVia #-}
+{-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
+{-# LANGUAGE NoTemplateHaskell #-}
+
 -- | "Data.Monoid" and "Data.Semigroup" re-exports, hiding conflicting symbols.
 module Himari.Prelude.Monoid
   ( module Export